backend/requirements.txt
Django~=5.0
django-ninja
psycopg2-binary
djangorestframework
djangorestframework-simplejwt
django-cors-headers
email-validator
